# common/api/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "FieldErrors": {
        "properties": {},
        "type": "object",
        "additionalProperties": {
          "properties": {
            "value": {},
            "message": {
              "type": "string"
            }
          },
          "required": ["message"],
          "type": "object"
        }
      },
      "Pick_ValidateError.status-or-message-or-fields_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/FieldErrors"
          }
        },
        "required": ["status", "message", "fields"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Pick_Error.message_": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      }
    }
  },
  "info": {
    "title": "api",
    "version": "0.1.0",
    "contact": {}
  },
  "paths": {
    "/project/{slug}": {
      "get": {
        "operationId": "GetProject",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "targetGroups": {
                      "items": {
                        "properties": {
                          "targetGroup": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["targetGroup"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "disciplines": {
                      "items": {
                        "properties": {
                          "discipline": {
                            "properties": {
                              "title": {
                                "type": "string"
                              }
                            },
                            "required": ["title"],
                            "type": "object"
                          }
                        },
                        "required": ["discipline"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "city": {
                      "type": "string"
                    },
                    "xing": {
                      "type": "string"
                    },
                    "instagram": {
                      "type": "string"
                    },
                    "youtube": {
                      "type": "string"
                    },
                    "twitter": {
                      "type": "string"
                    },
                    "linkedin": {
                      "type": "string"
                    },
                    "facebook": {
                      "type": "string"
                    },
                    "zipCode": {
                      "type": "string"
                    },
                    "streetNumber": {
                      "type": "string"
                    },
                    "street": {
                      "type": "string"
                    },
                    "website": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "excerpt": {
                      "type": "string"
                    },
                    "headline": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "background": {
                      "type": "string"
                    },
                    "logo": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "targetGroups",
                    "disciplines",
                    "city",
                    "xing",
                    "instagram",
                    "youtube",
                    "twitter",
                    "linkedin",
                    "facebook",
                    "zipCode",
                    "streetNumber",
                    "street",
                    "website",
                    "phone",
                    "email",
                    "description",
                    "excerpt",
                    "headline",
                    "name",
                    "id",
                    "background",
                    "logo",
                    "url"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "message": "Authentication failed",
                      "fields": {
                        "access_token": {
                          "message": "Invalid access token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pick_Error.message_"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": ["status"],
                      "type": "object"
                    }
                  ]
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 500,
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Retrieve a project by slug of the community including their public information.",
        "summary": "Retrieve project by slug.",
        "tags": ["Project"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}